fx-cache/index.html

<!DOCTYPE html>
<html lang="en">
<head>
  <meta charset="utf-8" />
  <title>Mocha</title>
  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
  <script crossorigin="anonymous" src='https://unpkg.com/mocha-chai-cdn/testharness/testharness.js'></script>
  <script crossorigin="anonymous" src='https://unpkg.com/mocha-chai-cdn/testharness/testharnessreport.js'></script>

</head>
<body>
  <div id="mocha"></div>
  <script type="module">
    import {renderFrames, chart, mkcanvas} from "../chart/dist/chart.js";
    import {linRamp, targets, sine, expRamp} from "./index.js";

    const l = mkcanvas();
    promise_test(async () => {
      const o = new ConvolverNode(new OfflineAudioContext(1, 0x1000, 0xffff), await linRamp(.1, 48000));

      const p = (await sine(new OfflineAudioContext(1, 0x1000, 0x1000), 19), [0, 1, 0.7, 0, 0, 0, 0, 2]).getChannelData(0);
      assert_true(true, "You win"); chart(mkcanvas(), p);// (await sine(new OfflineAudioContext(1, 0x1000, 0x1000))).getChannelData(0))
      console.log(p)
      const fl = await linRamp(.1, 48000).then(ab => ab.getChannelData(0));
      assert_true(fl.length == 4800)
      console.log(fl)
      chart(mkcanvas(), fl, 1, 4800);
      const fl2 = await targets(.1, 48000).then(ab => ab.getChannelData(0));
      assert_true(fl2.length == 4800)
      console.log(fl2)
      chart(mkcanvas(), fl2); assert_true(fl2.length == 4800)
    }, "freebie")
  </script>
</body>
</html>
